requirements.txt
selenium
pychrome
pydantic
lxml
python-dotenv
webdriver-manager
pygame
Flask
ReportLab
requests
psutil
lxml
beautifulsoup4